version 1.0.0: deploy; imporve docker and deploy using remote server; cd/ci with github actions

+++ src/test/java/at/backend/tourist/places/Service/ActivityServiceImplTest.java
@@ -77,7 +77,7 @@ public class ActivityServiceImplTest {
 
         assertNotNull(result);
         assertEquals("Hiking Tour", result.getName());
-        assertEquals(25.50, result.getPrice());
+        assertEquals(BigDecimal.valueOf(25.50), result.getPrice());
     }
 
     @Test
